+++ SerialPrograms/CMakeLists.txt
@@ -1455,6 +1455,8 @@ file(GLOB MAIN_SOURCES
     Source/PokemonSV/Programs/Farming/PokemonSV_LPFarmer.h
     Source/PokemonSV/Programs/Farming/PokemonSV_MaterialFarmer.cpp
     Source/PokemonSV/Programs/Farming/PokemonSV_MaterialFarmer.h    
+    Source/PokemonSV/Programs/Farming/PokemonSV_MaterialFarmerTools.cpp
+    Source/PokemonSV/Programs/Farming/PokemonSV_MaterialFarmerTools.h    
     Source/PokemonSV/Programs/Farming/PokemonSV_TournamentFarmer.cpp
     Source/PokemonSV/Programs/Farming/PokemonSV_TournamentFarmer.h
     Source/PokemonSV/Programs/Farming/PokemonSV_TournamentFarmer2.cpp
