+++ SerialPrograms/SerialPrograms.pro
@@ -726,6 +726,7 @@ SOURCES += \
     Source/PokemonSV/Programs/Farming/PokemonSV_GimmighoulRoamingFarmer.cpp \
     Source/PokemonSV/Programs/Farming/PokemonSV_LPFarmer.cpp \
     Source/PokemonSV/Programs/Farming/PokemonSV_MaterialFarmer.cpp \
+    Source/PokemonSV/Programs/Farming/PokemonSV_MaterialFarmerTools.cpp \
     Source/PokemonSV/Programs/Farming/PokemonSV_TournamentFarmer.cpp \
     Source/PokemonSV/Programs/Farming/PokemonSV_TournamentFarmer2.cpp \
     Source/PokemonSV/Programs/Farming/PokemonSV_WildItemFarmer.cpp \
@@ -1798,6 +1799,7 @@ HEADERS += \
     Source/PokemonSV/Programs/Farming/PokemonSV_GimmighoulRoamingFarmer.h \
     Source/PokemonSV/Programs/Farming/PokemonSV_LPFarmer.h \
     Source/PokemonSV/Programs/Farming/PokemonSV_MaterialFarmer.h \
+    Source/PokemonSV/Programs/Farming/PokemonSV_MaterialFarmerTools.h \
     Source/PokemonSV/Programs/Farming/PokemonSV_TournamentFarmer.h \
     Source/PokemonSV/Programs/Farming/PokemonSV_TournamentFarmer2.h \
     Source/PokemonSV/Programs/Farming/PokemonSV_WildItemFarmer.h \
