+++ SerialPrograms/Source/PokemonSV/Programs/Farming/PokemonSV_MaterialFarmer.h
@@ -16,6 +16,7 @@
 #include "NintendoSwitch/Options/NintendoSwitch_GoHomeWhenDoneOption.h"
 #include "PokemonSV/Options/PokemonSV_EncounterBotCommon.h"
 #include "PokemonSV/Options/PokemonSV_SandwichMakerOption.h"
+#include "PokemonSV_MaterialFarmerTools.h"
 
 namespace PokemonAutomation{
 namespace NintendoSwitch{
@@ -28,44 +29,26 @@ class MaterialFarmer_Descriptor : public SingleSwitchProgramDescriptor{
 public:
     MaterialFarmer_Descriptor();
 
-    struct Stats;
+    struct Stats : public LetsGoEncounterBotStats{
+        Stats()
+            : m_sandwiches(m_stats["Sandwiches"])
+            , m_autoheals(m_stats["Auto Heals"])
+            , m_game_resets(m_stats["Game Resets"])
+            , m_errors(m_stats["Errors"])
+        {
+            m_display_order.insert(m_display_order.begin() + 2, {"Sandwiches", HIDDEN_IF_ZERO});
+            m_display_order.insert(m_display_order.begin() + 3, {"Auto Heals", HIDDEN_IF_ZERO});
+            m_display_order.insert(m_display_order.begin() + 4, {"Game Resets", HIDDEN_IF_ZERO});
+            m_display_order.insert(m_display_order.begin() + 5, {"Errors", HIDDEN_IF_ZERO});
+        }
+        std::atomic<uint64_t>& m_sandwiches;
+        std::atomic<uint64_t>& m_autoheals;
+        std::atomic<uint64_t>& m_game_resets;
+        std::atomic<uint64_t>& m_errors;
+    };
     virtual std::unique_ptr<StatsTracker> make_stats() const override;
 };
 
-struct MaterialFarmerOptions {
-    BooleanCheckBoxOption& save_game_before_sandwich_option;
-    SimpleIntegerOption<uint16_t>& num_sandwich_rounds_option;
-    OCR::LanguageOCROption& language_option;
-    SandwichMakerOption& sandwich_options;
-    GoHomeWhenDoneOption& go_home_when_done_option;
-    FloatingPointOption& auto_heal_percent_option;
-    BooleanCheckBoxOption& save_debug_video_option;
-    BooleanCheckBoxOption& skip_warp_to_pokecenter_option;
-    BooleanCheckBoxOption& skip_sandwich_option;
-    EventNotificationOption& notification_status_update_option;
-    EventNotificationOption& notification_error_recoverable_option;
-    EventNotificationsOption& notifications_option;
-};
-
-void run_material_farmer(SingleSwitchProgramEnvironment& env, BotBaseContext& context, MaterialFarmerOptions options);
-
-void run_one_sandwich_iteration(SingleSwitchProgramEnvironment& env, BotBaseContext& context, 
-    LetsGoEncounterBotTracker& encounter_tracker, MaterialFarmerOptions& options);
-
-void move_to_start_position_for_letsgo0(SingleSwitchProgramEnvironment& env, BotBaseContext& context);
-
-void move_to_start_position_for_letsgo1(SingleSwitchProgramEnvironment& env, BotBaseContext& context);
-
-void lets_go_movement0(BotBaseContext& context);
-
-void lets_go_movement1(BotBaseContext& context);
-
-bool is_sandwich_expired(WallClock last_sandwich_time);
-
-void run_lets_go_iteration(SingleSwitchProgramEnvironment& env, BotBaseContext& context, LetsGoEncounterBotTracker& encounter_tracker);
-
-void run_from_battles_and_back_to_pokecenter(SingleSwitchProgramEnvironment& env, BotBaseContext& context, 
-    std::function<void(SingleSwitchProgramEnvironment& env, BotBaseContext& context)>&& action);
 
 class MaterialFarmer : public SingleSwitchProgramInstance{
 public:
